requirements.txt
google-adk[extensions]
fastapi
uvicorn[standard]
python-dotenv
httpx
python-nmap
dnspython
python-whois
